Add tree filtering to Advanced Import Settings * Implements node and resource filtering in the scene import settings dialog. * Moved TreeSortAndFilterBar class from ObjectDB Profiler module to the `editor/` folder so its available for every other Tree in the editor that has filtering.
Added
editor\gui\tree_sort_and_filter_line_edit.cpp
editor\gui\tree_sort_and_filter_line_edit.h

Updated
editor\import\3d\scene_import_settings.cpp
editor\import\3d\scene_import_settings.h
modules\objectdb_profiler\editor\data_viewers\class_view.cpp
modules\objectdb_profiler\editor\data_viewers\node_view.h
modules\objectdb_profiler\editor\data_viewers\object_view.h
modules\objectdb_profiler\editor\data_viewers\refcounted_view.h
modules\objectdb_profiler\editor\data_viewers\shared_controls.cpp
modules\objectdb_profiler\editor\data_viewers\shared_controls.h
