+++ modules/objectdb_profiler/editor/data_viewers/class_view.cpp
@@ -30,11 +30,10 @@
 
 #include "class_view.h"
 
-#include "shared_controls.h"
-
 #include "core/object/callable_mp.h"
 #include "core/object/class_db.h"
 #include "editor/editor_node.h"
+#include "editor/gui/tree_sort_and_filter_line_edit.h"
 #include "editor/themes/editor_scale.h"
 #include "scene/gui/split_container.h"
